+++ pacs/mdrun/analyzer/target.py
@@ -113,7 +113,7 @@ class Target(SuperAnalyzer):
         cmd_rmfile = f"rm {dir}/prd_image{extension}"
         subprocess.run(cmd_rmfile, shell=True)
 
-        rmsd_rep = np.loadtxt(f"{dir}/rms.xvg")[:, 1]
+        rmsd_rep = np.loadtxt(f"{dir}/rms.xvg", dtype="float32")[:, 1]
         return rmsd_rep
         # output of rms command will be tsv-like format
 
@@ -164,5 +164,5 @@ class Target(SuperAnalyzer):
             LOGGER.error(f"see {dir}/rms.log for details")
             exit(1)
 
-        rmsd = np.loadtxt(f"{dir}/rms.xvg")[:, 1]
+        rmsd = np.loadtxt(f"{dir}/rms.xvg", dtype="float32")[:, 1]
         return rmsd
